feat:  Remove LoadingLine component
Migrate LoadingLine component to cosmotech/ui
Adapt Loading view by using LoadingLine component from cosmotech/ui
Remove react-lottie from package.json
Update yarn.lock
