+++ package.json
@@ -28,7 +28,6 @@
     "react-dom": "^17.0.1",
     "react-fade-in": "^2.0.1",
     "react-i18next": "^11.9.0",
-    "react-lottie": "^1.2.3",
     "react-redux": "^7.2.3",
     "react-router-dom": "^5.2.0",
     "redux": "^4.0.5",
