--- src/components/LoadingLine/index.js
@@ -1,4 +1,0 @@
-// Copyright (c) Cosmo Tech.
-// Licensed under the MIT license.
-
-export { default } from './LoadingLine';
