+++ src/views/Loading/Loading.js
@@ -5,9 +5,12 @@ import React, { useEffect } from 'react';
 import PropTypes from 'prop-types';
 import Routes from '../../Routes';
 import FadeIn from 'react-fade-in';
-import LoadingLine from '../../components/LoadingLine';
+import { LoadingLine } from '@cosmotech/ui';
 import { STATUSES } from '../../state/commons/Constants';
 import { makeStyles } from '@material-ui/core';
+import * as dataLoading from '../../assets/loadingView/dataLoading.json';
+import * as dataLoaded from '../../assets/loadingView/dataLoaded.json';
+import * as dataError from '../../assets/loadingView/dataError.json';
 
 const useStyles = makeStyles(theme => ({
   panel: {
@@ -17,6 +20,32 @@ const useStyles = makeStyles(theme => ({
     transform: 'translate(-50%, -50%)'
   }
 }));
+
+const dataLoaderOptions = {
+  loop: true,
+  autoplay: true,
+  animationData: dataLoading.default,
+  rendererSettings: {
+    preserveAspectRatio: 'xMidYMid slice'
+  }
+};
+
+const dataLoadedOptions = {
+  loop: false,
+  autoplay: true,
+  animationData: dataLoaded.default,
+  rendererSettings: {
+    preserveAspectRatio: 'xMidYMid slice'
+  }
+};
+const errorDataOptions = {
+  loop: false,
+  autoplay: true,
+  animationData: dataError.default,
+  rendererSettings: {
+    preserveAspectRatio: 'xMidYMid slice'
+  }
+};
 
 const Loading = (
   {
@@ -55,10 +84,38 @@ const Loading = (
   return (authenticated && application.status !== STATUSES.SUCCESS
     ? (<div className={classes.panel} data-cy="loading-component">
           <FadeIn delay={200}>
-            <LoadingLine titleKey={'genericcomponent.loading.line.scenario.list.title'} hasError={hasErrors(scenarioList)} isLoading={isLoading(scenarioList)} height={120} width={120}/>
-            <LoadingLine titleKey={'genericcomponent.loading.line.dataset.list.title'} hasError={hasErrors(datasetList)} isLoading={isLoading(datasetList)} height={120} width={120}/>
-            <LoadingLine titleKey={'genericcomponent.loading.line.workspace.current.title'} hasError={hasErrors(workspace)} isLoading={isLoading(workspace)} height={120} width={120}/>
-              <LoadingLine titleKey={'genericcomponent.loading.line.solution.current.title'} hasError={hasErrors(solution)} isLoading={isLoading(solution)} height={120} width={120}/>
+            <LoadingLine titleKey={'genericcomponent.loading.line.scenario.list.title'}
+                         hasError={hasErrors(scenarioList)}
+                         isLoading={isLoading(scenarioList)}
+                         dataLoaderOptions={dataLoaderOptions}
+                         dataLoadedOptions={dataLoadedOptions}
+                         errorDataOptions={errorDataOptions}
+                         height={120}
+                         width={120}/>
+            <LoadingLine titleKey={'genericcomponent.loading.line.dataset.list.title'}
+                         hasError={hasErrors(datasetList)}
+                         isLoading={isLoading(datasetList)}
+                         dataLoaderOptions={dataLoaderOptions}
+                         dataLoadedOptions={dataLoadedOptions}
+                         errorDataOptions={errorDataOptions}
+                         height={120}
+                         width={120}/>
+            <LoadingLine titleKey={'genericcomponent.loading.line.workspace.current.title'}
+                         hasError={hasErrors(workspace)}
+                         isLoading={isLoading(workspace)}
+                         dataLoaderOptions={dataLoaderOptions}
+                         dataLoadedOptions={dataLoadedOptions}
+                         errorDataOptions={errorDataOptions}
+                         height={120}
+                         width={120}/>
+              <LoadingLine titleKey={'genericcomponent.loading.line.solution.current.title'}
+                           hasError={hasErrors(solution)}
+                           isLoading={isLoading(solution)}
+                           dataLoaderOptions={dataLoaderOptions}
+                           dataLoadedOptions={dataLoadedOptions}
+                           errorDataOptions={errorDataOptions}
+                           height={120}
+                           width={120}/>
           </FadeIn>
       </div>)
     : (<Routes authenticated={authenticated} authorized={authenticated} tabs={tabs}/>)
